Set permissions on media files to 0640
By default the permissions will be as they were for the Tempfile, which
is 0600.  This means that other users on the same repo won't be able to
get the media files.

This commit adds group read permissions to media files.  Group write
permissions are not required as the file is identified by its contents
anyway, so it should never need to be modified.

Conflicts:
	lib/git-media/filter-clean.rb